+++ CHANGELOG.md
@@ -12,6 +12,7 @@ the output from ```Spider.get_validated_metadata()``` regardless
 of whether or not it is specified as part of the explicit spider
 metadata declaration
 - added ```paranoia_level``` to spider metadata
+- added ```max_crawl_time_in_seconds``` to spider metadata
 
 ### Changed